feat: HoverActionProvider 추가
